+++ pandas/core/arrays/arrow/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa: F401
+
+from pandas.core.arrays.arrow.array import ArrowExtensionArray
